+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## Unreleased
+## 3.29.0
 
 - ref(nextjs): Adjust dev server command in verification message (#665)
+- feat(remix): Add feature selection (#646)
+- feat: Try running Prettier on exit (#644)
+- fix(remix): Don't create `.sentrclirc` if project uses Vite (#667)
 
 ## 3.28.0